Revert "feature: commission_rate_in_basis_points (#9822)" (#9847)
This reverts commit 416d4f67593bdd986dbb7f2b1b6a2ff7c0bfe920.

+++ src/flamenco/runtime/fd_runtime_stack.h
@@ -32,7 +32,7 @@ struct fd_vote_rewards {
   fd_pubkey_t pubkey;
   ulong       vote_rewards;
   uint        next;
-  ushort      commission;
+  uchar       commission;
 };
 typedef struct fd_vote_rewards fd_vote_rewards_t;
